+++ kairon/shared/chat/data_objects.py
@@ -2,7 +2,7 @@ from mongoengine import StringField, DictField, DateTimeField, ValidationError
 from datetime import datetime
 
 from kairon.shared.constants import ChannelTypes
-from kairon.shared.data.audit.base_data import Auditlog
+from kairon.shared.data.audit.data_objects import Auditlog
 from kairon.shared.data.signals import push_notification, auditlogger
 from kairon.shared.utils import Utility
