+++ kairon/shared/data/signals.py
@@ -88,5 +88,5 @@ def auditlogger_handler(event1):
 
 @auditlogger_handler(auditlog)
 def auditlogger(sender, document, **kwargs):
-    from kairon import Utility
-    Utility.save_and_publish_auditlog(document, sender.__name__, **kwargs)
+    from kairon.shared.data.audit.processor import AuditProcessor
+    AuditProcessor.save_and_publish_auditlog(document, sender.__name__, **kwargs)
