+++ tests/unit_test/api/api_processor_test.py
@@ -1,17 +1,18 @@
 import asyncio
 import datetime
 import os
+import time
 import uuid
 from unittest.mock import patch
 from urllib.parse import urljoin
 
 import jwt
+import pytest
 import responses
 from fastapi import HTTPException
 from fastapi_sso.sso.base import OpenID
 from mongoengine import connect
 from mongoengine.errors import ValidationError, DoesNotExist
-import pytest
 from mongomock.object_id import ObjectId
 from pydantic import SecretStr
 from pytest_httpx import HTTPXMock
@@ -20,12 +21,16 @@ from starlette.requests import Request
 from starlette.responses import RedirectResponse
 
 from kairon.api.app.routers.idp import get_idp_config
+from kairon.api.models import RegisterAccount, EventConfig, IDPConfig, StoryRequest, HttpActionParameters, Password
+from kairon.exceptions import AppException
+from kairon.idp.data_objects import IdpConfig
 from kairon.idp.processor import IDPProcessor
-from kairon.shared.admin.data_objects import BotSecrets
-from kairon.shared.auth import Authentication, LoginSSOFactory
 from kairon.shared.account.data_objects import Feedback, BotAccess, User, Bot, Account, Organization, TrustedDevice
 from kairon.shared.account.processor import AccountProcessor
+from kairon.shared.admin.data_objects import BotSecrets
+from kairon.shared.auth import Authentication, LoginSSOFactory
 from kairon.shared.authorization.processor import IntegrationProcessor
+from kairon.shared.data.audit.processor import AuditProcessor
 from kairon.shared.data.constant import ACTIVITY_STATUS, ACCESS_ROLES, TOKEN_TYPE, INTEGRATION_STATUS, \
     ORG_SETTINGS_MESSAGES, FeatureMappings
 from kairon.shared.data.data_objects import Configs, Rules, Responses
@@ -35,10 +40,6 @@ from kairon.shared.organization.processor import OrgProcessor
 from kairon.shared.sso.clients.facebook import FacebookSSO
 from kairon.shared.sso.clients.google import GoogleSSO
 from kairon.shared.utils import Utility, MailUtility
-from kairon.exceptions import AppException
-import time
-from kairon.idp.data_objects import IdpConfig
-from kairon.api.models import RegisterAccount, EventConfig, IDPConfig,StoryRequest, HttpActionParameters,Password
 
 os.environ["system_file"] = "./tests/testing_data/system.yaml"
 
@@ -934,7 +935,7 @@ class TestAccountProcessor:
         def _publish_auditlog(*args, **kwargs):
             return
 
-        monkeypatch.setattr(Utility, "save_and_publish_auditlog", _publish_auditlog)
+        monkeypatch.setattr(AuditProcessor, "save_and_publish_auditlog", _publish_auditlog)
         account = {
             "account": "Test_Account",
             "bot": "Test",
